ファイルの分離
